src/pytorch/hand_write3.py
import torch
import torch.nn as nn
import torch.optim as optim
from torchvision import datasets, transforms
from torch.utils.data import DataLoader
from PIL import Image


# 定义神经网络模型
class Net(nn.Module):
    def __init__(self):
        super(Net, self).__init__()
        self.fc1 = nn.Linear(28 * 28, 512)
        self.fc2 = nn.Linear(512, 256)
        self.fc3 = nn.Linear(256, 10)

    def forward(self, x):
        x = x.view(-1, 28 * 28)
        x = torch.relu(self.fc1(x))
        x = torch.relu(self.fc2(x))
        x = self.fc3(x)
        return x

    # 数据预处理


transform = transforms.Compose([
    transforms.ToTensor(),
    transforms.Normalize((0.5,), (0.5,))
])

# 加载数据集
trainset = datasets.MNIST('~/.pytorch/MNIST_data/', download=True, train=True, transform=transform)
trainloader = DataLoader(trainset, batch_size=64, shuffle=True)

testset = datasets.MNIST('~/.pytorch/MNIST_data/', download=True, train=False, transform=transform)
testloader = DataLoader(testset, batch_size=64, shuffle=False)

# 实例化模型、损失函数和优化器
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
model = Net().to(device)
criterion = nn.CrossEntropyLoss()
optimizer = optim.SGD(model.parameters(), lr=0.001, momentum=0.9)

# 训练模型
num_epochs = 5
for epoch in range(num_epochs):
    running_loss = 0.0
    for i, data in enumerate(trainloader, 0):
        inputs, labels = data[0].to(device), data[1].to(device)
        optimizer.zero_grad()
        outputs = model(inputs)
        loss = criterion(outputs, labels)
        loss.backward()
        optimizer.step()
        running_loss += loss.item()
    print(f'Epoch {epoch + 1}/{num_epochs}, Loss: {running_loss / len(trainloader)}')

print('Finished Training')

# 保存模型
torch.save(model.state_dict(), 'mnist_model.pth')
print('Model saved successfully!')


# 加载模型
def load_model(model_path, device):
    model = Net().to(device)
    model.load_state_dict(torch.load(model_path))
    model.eval()
    return model


loaded_model = load_model('mnist_model.pth', device)


# 使用加载的模型进行预测
def predict(image, model, transform):
    image = transform(image).unsqueeze(0).to(device)
    output = model(image)
    _, predicted_idx = torch.max(output, 1)
    return predicted_idx.item()


test_image_path = 'test_image.png'
test_image = Image.open(test_image_path).convert('L')
predicted_label = predict(test_image, loaded_model, transform)
print(f'Predicted label for the loaded model on image {test_image_path}: {predicted_label}')