requirements.txt
numpy~=1.26.4
torch~=2.2.2
torchvision~=0.17.2
tqdm~=4.66.2
pillow~=10.3.0
requests~=2.31.0